feat:ci-cd 설정 수정 및 패키지 이동에 따른 수정 (#366)
