+++ .github/workflows/ci-cd.yml
@@ -21,7 +21,7 @@ jobs:
   test:
     name: Run Tests
     runs-on: ubuntu-latest
-    
+
     services:
       postgres:
         image: postgres:16-alpine
@@ -36,7 +36,7 @@ jobs:
           --health-interval 10s
           --health-timeout 5s
           --health-retries 5
-      
+
       redis:
         image: redis:7-alpine
         ports:
@@ -46,7 +46,7 @@ jobs:
           --health-interval 10s
           --health-timeout 5s
           --health-retries 5
-      
+
       elasticsearch:
         image: docker.elastic.co/elasticsearch/elasticsearch:9.2.3
         env:
@@ -60,21 +60,21 @@ jobs:
           --health-interval 30s
           --health-timeout 10s
           --health-retries 5
-    
+
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
-      
+
       - name: Set up JDK 25
         uses: actions/setup-java@v4
         with:
           java-version: ${{ env.JAVA_VERSION }}
           distribution: 'temurin'
           cache: 'gradle'
-      
+
       - name: Grant execute permission for gradlew
         run: chmod +x ./fourtune/gradlew
-      
+
       - name: Run tests with Gradle
         working-directory: ./fourtune
         run: ./gradlew test --no-daemon
@@ -86,7 +86,7 @@ jobs:
           REDIS_HOST: localhost
           REDIS_PORT: 6379
           ELASTICSEARCH_URIS: http://localhost:9200
-      
+
       - name: Generate test report
         if: always()
         uses: dorny/test-reporter@v1
@@ -108,11 +108,11 @@ jobs:
     runs-on: ubuntu-latest
     needs: test
     if: github.event_name == 'push'
-    
+
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
-      
+
       - name: Set up JDK 25
         uses: actions/setup-java@v4
         with:
@@ -127,11 +127,11 @@ jobs:
 
       - name: Grant execute permission for gradlew
         run: chmod +x ./fourtune/gradlew
-      
+
       - name: Build with Gradle
         working-directory: ./fourtune
         run: ./gradlew build -x test --no-daemon
-      
+
       - name: Upload fourtune-api artifact
         uses: actions/upload-artifact@v4
         with:
@@ -146,16 +146,30 @@ jobs:
           path: fourtune/payment-service/build/libs/*.jar
           retention-days: 7
 
+      - name: Upload auction-service artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: auction-service-jar
+          path: fourtune/auction-service/build/libs/*.jar
+          retention-days: 7
+
+      - name: Upload recommendation-service artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: recommendation-service-jar
+          path: fourtune/recommendation-service/build/libs/*.jar
+          retention-days: 7
+
   docker-build:
     name: Build and Push Docker Image
     runs-on: ubuntu-latest
     needs: build
     if: github.event_name == 'push' && (github.ref == 'refs/heads/main' || github.ref == 'refs/heads/develop')
-    
+
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
-      
+
       - name: Download fourtune-api artifact
         uses: actions/download-artifact@v4
         with:
@@ -167,6 +181,18 @@ jobs:
         with:
           name: payment-service-jar
           path: fourtune/payment-service/build/libs
+
+      - name: Download auction-service artifact
+        uses: actions/download-artifact@v4
+        with:
+          name: auction-service-jar
+          path: fourtune/auction-service/build/libs
+
+      - name: Download recommendation-service artifact
+        uses: actions/download-artifact@v4
+        with:
+          name: recommendation-service-jar
+          path: fourtune/recommendation-service/build/libs
 
       - name: Make Firebase Key File for Docker
         run: |
@@ -190,7 +216,9 @@ jobs:
           push: true
           build-args: |
             MODULE_NAME=fourtune-api
-          tags: ghcr.io/${{ github.repository_owner }}/fourtune-api:${{ github.sha }}
+          tags: |
+            ghcr.io/${{ github.repository_owner }}/fourtune:latest
+            ghcr.io/${{ github.repository_owner }}/fourtune:${{ github.sha }}
           cache-from: type=gha
           cache-to: type=gha,mode=max
 
@@ -201,7 +229,35 @@ jobs:
           push: true
           build-args: |
             MODULE_NAME=payment-service
-          tags: ghcr.io/${{ github.repository_owner }}/payment-service:${{ github.sha }}
+          tags: |
+            ghcr.io/${{ github.repository_owner }}/payment-service:latest
+            ghcr.io/${{ github.repository_owner }}/payment-service:${{ github.sha }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+
+      - name: Build and push auction-service image
+        uses: docker/build-push-action@v5
+        with:
+          context: ./fourtune
+          push: true
+          build-args: |
+            MODULE_NAME=auction-service
+          tags: |
+            ghcr.io/${{ github.repository_owner }}/auction-service:latest
+            ghcr.io/${{ github.repository_owner }}/auction-service:${{ github.sha }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+
+      - name: Build and push recommendation-service image
+        uses: docker/build-push-action@v5
+        with:
+          context: ./fourtune
+          push: true
+          build-args: |
+            MODULE_NAME=recommendation-service
+          tags: |
+            ghcr.io/${{ github.repository_owner }}/recommendation-service:latest
+            ghcr.io/${{ github.repository_owner }}/recommendation-service:${{ github.sha }}
           cache-from: type=gha
           cache-to: type=gha,mode=max
 
@@ -213,7 +269,7 @@ jobs:
     environment:
       name: development
       url: https://dev.fourtune.com
-    
+
     steps:
       - name: Deploy to development server
         uses: appleboy/ssh-action@v1.0.0
@@ -238,7 +294,7 @@ jobs:
     environment:
       name: production
       url: https://api.fourtune.store
-    
+
     steps:
       - name: Deploy to production server
         uses: appleboy/ssh-action@v1.0.0
@@ -255,13 +311,6 @@ jobs:
             docker compose -f docker-compose.prod.yml pull
             docker compose -f docker-compose.prod.yml up -d --no-deps
             docker system prune -f
-            # 4. 내부 헬스 체크 (외부 https가 막혀있으므로 localhost 사용)
             echo "Waiting for service to start..."
             sleep 30
             curl -f http://localhost:8080/actuator/health || echo "Warning: Health check failed but deployment finished."
-#
-#      - name: Health check
-#        run: |
-#          sleep 30
-#          curl -f ${{ secrets.PROD_URL }}/actuator/health || exit 1
-
