+++ fourtune/payment-service/src/main/java/com/fourtune/payment/infrastructure/kafka/settlement/SettlementOutboxEventHandler.java
@@ -1,4 +1,4 @@
-package com.fourtune.payment.infrastructure.kafka;
+package com.fourtune.payment.infrastructure.kafka.settlement;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fourtune.outbox.handler.OutboxEventHandler;
@@ -8,7 +8,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
 /**
- * 결제 도메인 Outbox 이벤트 핸들러
+ * 정산 도메인 Outbox 이벤트 핸들러
  * payload = {"eventType":"...","aggregateId":123,"data":{...}} 형태로 저장된 JSON을 파싱
  * 후 Kafka 발행
  */
@@ -16,11 +16,11 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 @ConditionalOnProperty(name = "feature.kafka.enabled", havingValue = "true", matchIfMissing = false)
-public class PaymentOutboxEventHandler implements OutboxEventHandler {
+public class SettlementOutboxEventHandler implements OutboxEventHandler {
 
-    private static final String AGGREGATE_TYPE = "Payment";
+    private static final String AGGREGATE_TYPE = "Settlement";
 
-    private final PaymentKafkaProducer paymentKafkaProducer;
+    private final SettlementKafkaProducer settlementKafkaProducer;
     private final ObjectMapper objectMapper;
 
     @Override
@@ -30,10 +30,10 @@ public class PaymentOutboxEventHandler implements OutboxEventHandler {
 
     @Override
     public void handle(String payload) throws Exception {
-        PaymentEventPayload wrapper = objectMapper.readValue(payload, PaymentEventPayload.class);
+        SettlementEventPayload wrapper = objectMapper.readValue(payload, SettlementEventPayload.class);
         String key = String.valueOf(wrapper.getAggregateId());
         String eventType = wrapper.getEventType();
         String value = objectMapper.writeValueAsString(wrapper.getData());
-        paymentKafkaProducer.sendSync(key, value, eventType);
+        settlementKafkaProducer.sendSync(key, value, eventType);
     }
 }
